Fix applied to display an error when non-owner tries to edit or delete the images'

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
@@ -91,14 +92,27 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
+    //Fix applied to have only owner able to edit the images
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session, RedirectAttributes redirectAttributes) {
         Image image = imageService.getImage(imageId);
+        Integer ownerId = image.getUser().getId();
+        User user = (User) session.getAttribute("loggeduser");
+        Integer loggedUser = user.getId();
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if(loggedUser == ownerId) {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+            String error = "Only the owner of the image can edit the image";
+            //Need to use redirectAttributes to show the error message on the page
+            //Adding messages to model does not display the error messages once the page is redirected
+            redirectAttributes.addFlashAttribute("editError", error);
+            return "redirect:/images/"+ image.getId() +"/"+ image.getTitle();
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,17 +146,32 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/"+ updatedImage.getId() + "/" + updatedImage.getTitle();
     }
 
 
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+    //Fix applied to have only owner able to delete the images
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirectAttributes) {
+        Image image = imageService.getImage(imageId);
+        Integer ownerId = image.getUser().getId();
+        User user = (User) session.getAttribute("loggeduser");
+        Integer loggedUser = user.getId();
+
+        if(loggedUser == ownerId) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            String error = "Only the owner of the image can delete the image";
+            //Need to use redirectAttributes to show the error message on the page
+            //Adding messages to model does not display the error messages once the page is redirected
+            redirectAttributes.addFlashAttribute("deleteError", error);
+            return "redirect:/images/"+ image.getId() +"/"+ image.getTitle();
+        }
     }
